Moved and deleted unnecessary dial_switch files now. Looking to make a new subclass of the display US2066 class to add some functions for the switch cases rather than have it messy in the dial switch class.
