--- testing/HelloWorld_i2c/HelloWorld_i2c.ino
@@ -1,51 +1,0 @@
-/*
- Demonstration sketch for Adafruit i2c/SPI LCD backpack
- using MCP23008 I2C expander
- ( https://learn.adafruit.com/i2c-spi-lcd-backpack )
-
- This sketch prints "Hello World!" to the LCD
- and shows the time.
-
-  The circuit:
- * 5V to Arduino 5V pin
- * GND to Arduino GND pin
- * CLK to Analog #5
- * DAT to Analog #4
-*/
-
-// include the library code:
-#include "Adafruit_LiquidCrystal.h"
-
-// Connect via i2c, default address #0 (A0-A2 not jumpered)
-Adafruit_LiquidCrystal lcd(0);
-
-void setup() {
-  Serial.begin(115200);
-  // while(!Serial);
-  Serial.println("LCD Character Backpack I2C Test.");
-
-  // set up the LCD's number of rows and columns:
-  // NOTE: can be tested on our 20 x 4 display at Q-hut. Change (16, 2) -> (20, 4)
-  if (!lcd.begin(16, 2)) { 
-    Serial.println("Could not init backpack. Check wiring.");
-    while(1);
-  }
-  Serial.println("Backpack init'd.");
-
-  // Print a message to the LCD.
-  lcd.print("hello, world!");
-}
-
-void loop() {
-  // set the cursor to column 0, line 1
-  // (note: line 1 is the second row, since counting begins with 0):
-  lcd.setCursor(0, 1);
-  // print the number of seconds since reset:
-  lcd.print(millis()/1000);
-
-  lcd.setBacklight(HIGH);
-  delay(500);
-  lcd.setBacklight(LOW);
-  delay(500);
-}
-
