+++ testing/display_menu/display_menu.ino
@@ -1,35 +1,36 @@
+/*
+  Modified code of ReadAnalogVoltage example and dial_switch_case_concept.md
+
+  Kinda jank but enough for testing the dial switches
+
+  Wanted to put in a redundancy check for the for loop in get_dial_pos() function
+  but requires global vars and wasn't sure how to do that atm. Leaving as a TODO task
+*/
 #include "dial_switch.h"
-US2066 OLED;
+const int DRIVE_DIAL = A0;
+const int MENU_DIAL = A1;
+Dial_Switch drive_dial(DRIVE_DIAL_PIN);
+Dial_Switch menu_dial(MENU_DIAL_PIN);
 
-int DRIVE_DIAL = A0;
-int MENU_DIAL = A1;
-
-unsigned char wheelSpeed[8];
-
-
-int menu_pos_buff;
 
 void setup() {
+  // initialize serial communication at 9600 bits per second:
   Serial.begin(9600);
-  Wire.begin();
-  OLED.init();
 }
 
 void loop() {
-  // Get the dial position (number)
-  int drive_pos = get_dial_pos(DRIVE_DIAL);
-  int menu_pos = get_dial_pos(MENU_DIAL);
+  // Read in the dial position (integer)
+  drive_dial.read_adc();
+  menu_dial.read_adc();
 
-  if(!drive_pos || !menu_pos) {
+  // Serial.println(drive_pos);
+  if(!drive_dial.get_dial_pos() || !menu_dial.get_dial_pos()) {
     Serial.println("Dial Switch Error");
+    return 0;
   }
 
-  // clear when switching menus
-  if(menu_pos != menu_pos_buff) {
-    OLED.clear();
-  }
-  menu_pos_buff = menu_pos;
-  menu_switch_case(menu_pos, OLED);
-  // drive_switch_case(drive_pos, OLED);
+  // Use dial position for switch cases
+  drive_switch_case(drive_pos);
+  menu_switch_case(menu_pos);
 
 }
